requirements.txt
openai
getPDF
requests
bs4
pillow